Update lineage_enchilada.mk
Model is inaccurate anyways, some models are A6000 and others A6003 but it always say A6003 so no reason to put that instead let's put 'OnePlus 6' so atleast it gets identified as a 'OnePlus 6' in apps (many don't, for example accubattery and Google). Expecting a toxic response but I thought this was a good idea.